firebase + nyoba bikin register
